+++ server/routes/status.js
@@ -43,11 +43,21 @@ export default (app) => {
 
   app.delete('/statuses/:id', { name: 'deleteStatus', preHandler: app.auth([app.authCheck]) }, async (request, reply) => {
     try {
-      await app.objection.models.status.query().deleteById(request.params.id);
-      request.flash('success', i18next.t('views.pages.statuses.delete.success'));
+      const relatedStatuses = await app.objection.models.task
+        .query()
+        .select()
+        .where('status_id', '=', request.params.id);
+      if (relatedStatuses.length > 0) {
+        request.flash('error', i18next.t('views.pages.statuses.delete.errUsed'));
+      } else {
+        await app.objection.models.status.query().deleteById(request.params.id);
+        request.flash('success', i18next.t('views.pages.statuses.delete.success'));
+      }
       reply.redirect(app.reverse('statuses'));
-    } catch {
+    } catch (e) {
+      request.log.error(e);
       request.flash('error', i18next.t('views.pages.statuses.delete.error'));
+      reply.redirect(app.reverse('statuses'));
     }
   });
